pyproject.toml
[tool.poetry]
name = "p"
version = "0.1.0"
description = ""
authors = ["mohser sh <sh1361m@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.11.5"
mlflow = "^2.6.0"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
prefect = "^2.12.0"
prefect-ray = "^0.2.5"
grpcio = "^1.57.0"
ray = "^2.6.3"
pytest = "^7.4.1"
boto3 = "^1.28.40"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
